# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "myhome-backend",
    "description": "Aplicación para Desarrollo de Aplicaciones 1 - Only in vercel",
    "version": "1.0.0"
  },
  "host": "myhome-backend.vercel.app",
  "basePath": "/api/v1/",
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/real-state-companies/properties": {
      "get": {
        "tags": [
          "Real State Companies"
        ],
        "description": "Obtain properties for a real state company.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "Auth API.",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "query",
            "description": "Real State Company ID for get own properties.",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}